% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrivers.R
\name{calcGeneMutScores}
\alias{calcGeneMutScores}
\title{Score mutations for each gene in whitelist}
\usage{
calcGeneMutScores(
  driver.catalog.path = NULL,
  df = NULL,
  gene.whitelist.path = GENE_MUT_WHITELIST,
  tag.features = T
)
}
\arguments{
\item{driver.catalog.path}{Path to LINX driver catalog txt file}

\item{df}{A dataframe of the driver catalog txt file}

\item{gene.whitelist.path}{Path to txt file with list of selected genes}

\item{tag.features}{If TRUE, will tag the value names with the feature type}
}
\value{
An integer vector
}
\description{
Score mutations for each gene in whitelist
}
