% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{combineFeatures}
\alias{combineFeatures}
\title{Combine feature columns by summing them together}
\usage{
combineFeatures(x, target.features = NULL, regex = NULL, target.name = NULL)
}
\arguments{
\item{x}{A matrix or dataframe}

\item{target.features}{A character vector of column (feature) names}

\item{regex}{Instead of target.features, a regex can be specified}

\item{target.name}{A name to assign the new feature. If unspecified, will use the name of the
first old feature}
}
\value{
The original matrix or dataframe with the indicated features combined
}
\description{
Combine feature columns by summing them together
}
