% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractContextsSvLinx.R
\name{countBinnedSvLen}
\alias{countBinnedSvLen}
\title{Count binned SV length}
\usage{
countBinnedSvLen(v.sv.len, bins = c(0, 10^c(3:7), Inf))
}
\arguments{
\item{v.sv.len}{A numeric vector of SV lengths}

\item{bins}{A numeric vector specifying the interval breaks}
}
\value{
An integer vector of counts
}
\description{
Count binned SV length
}
