% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectKataegis.R
\name{detectKataegis}
\alias{detectKataegis}
\title{Detect kataegis from SNVs}
\usage{
detectKataegis(
  vcf.file = NULL,
  df = NULL,
  vcf.filter = "PASS",
  keep.chroms = c(1:22, "X"),
  ref.genome = mutSigExtractor::DEFAULT_GENOME,
  merge.consecutive = F,
  min.muts = 6,
  imd.thres = 1000,
  cp = 1e-06,
  max.imd.quantile = 1,
  output.type = 0,
  verbose = T,
  ...
)
}
\arguments{
\item{vcf.file}{Path to the vcf file containing SNVs}

\item{df}{A bed file-like dataframe with the columns: chrom, pos, ref, alt}

\item{vcf.filter}{Which variants to filter for? Default is 'PASS'. Only applies if vcf.file is
specified}

\item{keep.chroms}{Which chromosomes to keep? Default is c(1:22,'X')}

\item{ref.genome}{A BSgenome reference genome. Default is BSgenome.Hsapiens.UCSC.hg19. If another
reference genome is indicated, it will also need to be installed.}

\item{merge.consecutive}{Some vcfs report MNVs as consecutive variants. For these vcfs, such rows
need to be merged into one row for proper function of downstream mutSigExtractor functions.}

\item{min.muts}{Min number of SNVs in a PCF segment for the segment to be considered a kataegis
focus}

\item{imd.thres}{Only PCF segments with mean intermutation distance less that this value will be
considered kataegis foci}

\item{cp}{rpart.control complexity parameter. Higher values increase overfitting of PCF}

\item{max.imd.quantile}{The intermutation distance at this quantile is calculated. If
max.imd.quantile < imd.thres, imd.thres will be set to max.imd.quantile}

\item{output.type}{0: simplified data of kataegis segments (chrom and pos refer to the first
variant in the kataegis focus). 1: data from all kataegis associated mutations, 2: raw data. Note
that the last variant from each chromosome is not reported.}

\item{verbose}{}

\item{...}{}
}
\value{
A dataframe
}
\description{
Detects kataegis from SNVs provided from a vcf file or dataframe. A piecewise
constant fit (using rpart (R in built regression tree algorithm)) is performed on the
intermutation distances. By default, segments with min.muts>=6 and imd.thres<=1000 are considered
kataegis foci. These default values were chosen according to d'Antonio et al Cell Rep. 2016
(https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4972030/)
}
