% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectKataegis.R
\name{extractContextsKataegis}
\alias{extractContextsKataegis}
\title{Make kataegis contexts}
\usage{
extractContextsKataegis(
  vcf.snv,
  vcf.sv,
  vcf.filter = "PASS",
  ref.genome = mutSigExtractor::DEFAULT_GENOME,
  keep.chroms = c(1:22, "X"),
  merge.consecutive = F,
  kat.foci.flank.length = 1000,
  verbose = T
)
}
\arguments{
\item{vcf.snv}{Path to the vcf file containing SNVs}

\item{vcf.sv}{Path to the GRIDSS vcf file containing}

\item{vcf.filter}{Which variants to filter for? Default is 'PASS'. Only applies if vcf.file is
specified}

\item{ref.genome}{A BSgenome reference genome. Default is BSgenome.Hsapiens.UCSC.hg19. If another
reference genome is indicated, it will also need to be installed.}

\item{keep.chroms}{Which chromosomes to keep? Default is c(1:22,'X')}

\item{merge.consecutive}{Some vcfs report MNVs as consecutive variants. For these vcfs, such rows}

\item{kat.foci.flank.length}{Number of bp on either side of kataegis to scan for SVs}

\item{verbose}{Show progress messages?}
}
\value{
An integer vector
}
\description{
Make kataegis contexts
}
