% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractContextsMnv.R
\name{extractContextsMnv}
\alias{extractContextsMnv}
\title{Extract MNV contexts}
\usage{
extractContextsMnv(
  vcf.file = NULL,
  df = NULL,
  vcf.filter = "PASS",
  keep.chroms = c(1:22, "X"),
  len.cap = 5,
  ref.genome = mutSigExtractor::DEFAULT_GENOME,
  tag.features = T,
  verbose = F,
  ...
)
}
\arguments{
\item{vcf.file}{Path to the vcf file}

\item{df}{A dataframe containing the columns: chrom, pos, ref, alt. Alternative input option to
vcf.file}

\item{vcf.filter}{A character or character vector to specifying which variants to keep,
corresponding to the values in the vcf FILTER column}

\item{keep.chroms}{A character vector specifying which chromosomes to keep (chromosome names
should be in the style of the vcf). To keep autosomal and sex chromosomes for example use:
keep.chroms=c(1:22,'X','Y')}

\item{len.cap}{MNVs longer than this length will be put into the same bin. MNV length is defined
as max(nchar(REF),nchar(ALT))}

\item{ref.genome}{A BSgenome reference genome. Default is BSgenome.Hsapiens.UCSC.hg19. If another
reference genome is indicated, it will also need to be installed.}

\item{tag.features}{If TRUE, will tag the value names with the feature type}

\item{verbose}{Print progress messages?}

\item{...}{Other arguments that can be passed to mutSigExtractor::variantsFromVcf()}
}
\value{
An integer vector
}
\description{
Extract MNV contexts
}
