% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRmdSnv.R
\name{extractRmdSnv}
\alias{extractRmdSnv}
\title{Calculated regional mutation density (rmd)}
\usage{
extractRmdSnv(
  vcf.file,
  ref.genome = mutSigExtractor::DEFAULT_GENOME,
  verbose = F,
  sample.name = NULL,
  bin.size = 1e+06,
  as.matrix = T,
  ...
)
}
\arguments{
\item{vcf.file}{Path to the vcf file}

\item{verbose}{Print progress messages?}

\item{sample.name}{If a character is provided, the header for the output
matrix will be named to this. If none is provided, the basename of the vcf
file will be used.}

\item{bin.size}{Size of genomic bins}

\item{as.matrix}{Return a matrix or vector?}

\item{...}{Other arguments that can be passed to variantsFromVcf()}

\item{ref.genomeA}{BSgenome reference genome. Default is
BSgenome.Hsapiens.UCSC.hg19. If another reference genome is indicated, it
will also need to be installed.}
}
\value{
A matrix or vector
}
\description{
Calculated regional mutation density (rmd)
}
