% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrivers.R
\name{getGeneAmpLevels}
\alias{getGeneAmpLevels}
\title{Get minCopyNumber for set of genes}
\usage{
getGeneAmpLevels(
  driver.catalog.path = NULL,
  df = NULL,
  gene.whitelist.path = GENE_AMP_WHITELIST,
  tag.features = T
)
}
\arguments{
\item{driver.catalog.path}{Path to LINX driver catalog txt file}

\item{df}{A dataframe of the driver catalog txt file}

\item{gene.whitelist.path}{Path to txt file with list of selected genes}

\item{tag.features}{If TRUE, will tag the value names with the feature type}
}
\value{
A numeric vector
}
\description{
Get minCopyNumber for set of genes
}
