% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrivers.R
\name{getGeneMutStatus}
\alias{getGeneMutStatus}
\title{Get mutation status from a whitelist of genes}
\usage{
getGeneMutStatus(
  driver.catalog.path = NULL,
  df = NULL,
  gene.whitelist.path,
  mode,
  tag.features = T
)
}
\arguments{
\item{driver.catalog.path}{Path to LINX driver catalog txt file}

\item{df}{A dataframe of the driver catalog txt file}

\item{gene.whitelist.path}{Path to txt file with list of selected genes}

\item{mode}{Can be 'biallelic', 'monoallel', or 'amp'}

\item{tag.features}{If TRUE, will tag the value names with the feature type}
}
\value{
A (1,0 vector)
}
\description{
Get mutation status from a whitelist of genes
}
