% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPurplePurityData.R
\name{getPurplePurityData}
\alias{getPurplePurityData}
\title{Get gender and whole genome duplication status}
\usage{
getPurplePurityData(purple.purity.path, add.feature.type = F)
}
\arguments{
\item{purple.purity.path}{Path to purple.purity.tsv file}

\item{add.feature.type}{If TRUE, will tag the value names with the feature
type}
}
\value{
A dataframe
}
\description{
Get gender and whole genome duplication status
}
