% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getViralInsertions.R
\name{getViralInsertions}
\alias{getViralInsertions}
\title{Determines the presence of viral inserts in a sample}
\usage{
getViralInsertions(
  linx.viral.inserts.path,
  viral.host.ref.path = VIRAL_HOST_REF,
  return.counts = F,
  tag.features = T
)
}
\arguments{
\item{linx.viral.inserts.path}{Path to txt file with the columns: SampleId,
SvId, VirusId, VirusName}

\item{viral.host.ref.path}{Path to txt file with the columns: refseq_id,
virus_name, feature_group, cancer_associated}

\item{return.counts}{If TRUE, returns the actual number of viral inserts
found in a sample}

\item{tag.features}{If TRUE, will tag the value names with the feature
type}
}
\value{
An integer vector
}
\description{
Determines the presence of viral inserts in a sample
}
