% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{linearizeChromPos}
\alias{linearizeChromPos}
\title{Converts chrom/pos into linear coordinates}
\usage{
linearizeChromPos(chrom, pos, ref.genome = mutSigExtractor::DEFAULT_GENOME)
}
\arguments{
\item{chrom}{A character vector indicating the chromosome}

\item{pos}{An integer vector indicating the position}

\item{ref.genome}{A BSgenome object}
}
\value{
An integer vector of linear coordinates
}
\description{
Converts chrom/pos into linear coordinates
}
