% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrivers.R
\name{scoreDriverCatalog}
\alias{scoreDriverCatalog}
\title{Calculate score for LINX driver mutation types}
\usage{
scoreDriverCatalog(df)
}
\arguments{
\item{df}{A dataframe of the driver catalog txt file}
}
\value{
An integer vector
}
\description{
Calculate score for LINX driver mutation types
}
\examples{
df <- read.delim('/Users/lnguyen/hpc/cuppen/projects/P0013_WGS_patterns_Diagn/CUPs_classifier/processed/cuplr/hmfSvFeatureExtractor/scripts/select_drivers//driver.catalog.merged.txt.gz')
scoreDriverCatalog(df)
}
