% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitFeatures.R
\name{splitDfByFeatureGroup}
\alias{splitDfByFeatureGroup}
\title{Split a dataframe into feature groups}
\usage{
splitDfByFeatureGroup(df, rm.tags = F)
}
\arguments{
\item{df}{A dataframe where colnames are in the form
'feature_tag.feature_name'}

\item{rm.tags}{Remove feature group tags?}
}
\value{
A list of dataframes
}
\description{
Split a dataframe into feature groups
}
